Revise collection tests
This commit should result in several failing tests that reveal bugs.

Tests for collections are now consistently structured in three lines:
  1. Test positive case with subscripting
  2. Test negative case with correct outer type but wrong inner type
  3. Test negative case with wrong outer type

In addition, I chose simpler test values with empty collections so the
tests may be read and understood faster. The new values more directly
test the key properties of the collections and I tried to reuse literals
so the tests could be read and understood faster. Using an empty string
has the added benefit of surfacing some edge cases.