Revise collection tests
This commit should result in several failing tests that reveal bugs.

Consistency in test cases was improved and most tests are now structured
as follows where applicable:
  1. Test positive case with subscripting
  2. Test positive case with nested subscripting
  3. Test negative case with correct outer type but wrong inner type
  4. Test negative case with wrong outer type

In addition, I chose simpler test values with empty collections so the
tests may be read and understood faster. The new values more directly
test the key properties of the collections and I tried to reuse literals
so the tests could be read and understood faster. Moreover, I adjusted
the test values to cover more edge cases.

For instance, when the object...

- has a `len` of 0;
- is a valid instance of the type it is checked against;
- is of a type that itself is not subscriptable;
- is of a type that is a subclass of a type that is subscriptable;

...and it is checked against a type that is subscriptable, a bug may be
observed.

Maybe there is a simpler way of defining this, but it's hard to think of
examples where the object is not `''`.

All tests are passing now, but `iter('')` is still unsupported.